benchmarks/sims/requirements.txt
beautifulsoup4==4.13.4
low-index==1.2.1
lxml==5.4.0
pytest-benchmark==5.1.0
rich==14.0.0
